# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MAX Together Backend API
  version: 1.0.0-mvp
  description: Финальный контракт MVP для персонального поиска событий и совместного выбора в комнате из двух участников в
    MAX.
servers:
- url: https://worknet.team/api/v1
  description: Production
- url: http://localhost:8080/api/v1
  description: Local
...
security:
- bearerAuth: []
paths:
...
  /events/{eventId}/ticket-click:
    post:
      tags:
      - Discovery
      operationId: recordTicketClick
      summary: Зафиксировать переход к билетам и получить безопасный URL
      parameters:
      - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketClickRequest'
            example:
              source: event_detail
              room_id: null
      responses:
        '200':
          description: Проверенный внешний HTTPS URL разрешённого ticket-провайдера; приложение открывает его через MAX Bridge
            или браузер с безопасным режимом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketClickResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session
      description: Случайный bearer token сессии MAX, действующий 24 часа; хранится на сервере только в виде hash.
  parameters:
...
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthenticated:
      description: Token отсутствует, истёк или невалиден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не существует или недоступен пользователю
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Конфликт состояния ресурса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    TicketClickRequest:
      type: object
      additionalProperties: false
      required:
      - source
      properties:
        source:
          type: string
          enum:
          - event_detail
          - match
          - saved
          - feed
        room_id:
          type:
          - string
          - 'null'
          format: uuid
    TicketClickResponse:
      type: object
      required:
      - external_url
      properties:
        external_url:
          type: string
          format: uri
          description: Внешний HTTPS URL провайдера из allowlist; не является внутренним redirect или временным токеном
...
    RoomState:
      type: string
      enum:
      - collecting_intents
      - ranking
      - voting
      - matched
      - exhausted
    RoomSummary:
      type: object
      required:
      - id
      - name
      - city_id
      - state
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        city_id:
          type: string
          format: uuid
        state:
          $ref: '#/components/schemas/RoomState'
    PublicParticipant:
      type: object
      required:
      - id
      - display_name
      - role
      - intent_ready
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type:
          - string
          - 'null'
          format: uri
        role:
          type: string
          enum:
          - creator
          - participant
        intent_ready:
          type: boolean
          description: Подтверждение intent для текущего round; не раскрывает содержимое intent
...
    MatchSummary:
      type: object
      required:
      - id
      - room_id
      - event_id
      - matched_at
      - participants
      properties:
        id:
          type: string
          format: uuid
        room_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        matched_at:
          type: string
          format: date-time
        participants:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: '#/components/schemas/PublicParticipant'
          description: Только безопасные display_name/avatar/role; MAX ID и intent не раскрываются
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
              enum:
              - UNAUTHENTICATED
              - TOKEN_EXPIRED
              - FORBIDDEN
              - VALIDATION_FAILED
              - NOT_FOUND
              - CONFLICT
              - IDEMPOTENCY_CONFLICT
              - ACTIVE_ROOM_EXISTS
              - RATE_LIMITED
              - ROOM_FULL
              - INVITE_EXPIRED
              - ROOM_EXPIRED
              - ROOM_NOT_READY
              - POOL_NOT_READY
              - POOL_EXHAUSTED
              - STALE_POOL_VERSION
              - INTENT_LOCKED
              - VOTE_ALREADY_CAST
              - ALREADY_MATCHED
              - ROUND_LIMIT_REACHED
              - EVENT_UNAVAILABLE
              - TICKETS_UNAVAILABLE
              - INTERNAL
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
              properties:
                current_pool_version:
                  type: integer
                  minimum: 1
                my_pool_finished:
                  type: boolean
                room_exhausted:
                  type: boolean
                field:
                  type: string
                fields:
                  type: array
                  items:
                    type: object
                    required:
                    - field
                    - reason
                    properties:
                      field:
                        type: string
                      reason:
                        type: string
                active_room:
                  anyOf:
                  - $ref: '#/components/schemas/RoomSummary'
                  - type: 'null'
                match:
                  anyOf:
                  - $ref: '#/components/schemas/MatchSummary'
                  - type: 'null'